requirements.txt
Django==4.2.1
pillow==9.5.0
djlint
djangorestframework==3.14.0
markdown
django-filter
psycopg2-binary==2.9.6